+++ src/components/Candidates/index.tsx
@@ -6,20 +6,29 @@ const Candidates = (candidates: IfcCandidates) => {
     candidateImageSources,
     candidatesData,
     popularVoteTotals,
+    renderPropErrorMessage,
     winnerTakeAllTotals,
   } = candidates;
-  const candidatesMap = candidatesData.map((candidate, i) => (
-    <Candidate
-      imageSrc={
-        candidateImageSources?.[candidate.party as keyof CandidateImageSrc]
-      }
-      name={candidate.name}
-      key={`candidate-${i}`}
-      party={candidate.party}
-      popularVoteTotals={popularVoteTotals[i]}
-      winnerTakeAllTotal={winnerTakeAllTotals[i]}
-    />
-  ));
+  const candidatesMap = candidatesData.map((candidate, i) => {
+    if (!candidate.name || !candidate.party) {
+      renderPropErrorMessage(
+        'candidatesData',
+        'candidatesData prop is malformed'
+      );
+    }
+    return (
+      <Candidate
+        imageSrc={
+          candidateImageSources?.[candidate.party as keyof CandidateImageSrc]
+        }
+        name={candidate.name}
+        key={`candidate-${i}`}
+        party={candidate.party}
+        popularVoteTotals={popularVoteTotals[i]}
+        winnerTakeAllTotal={winnerTakeAllTotals[i]}
+      />
+    );
+  });
   return <div className="candidatesWrapper">{candidatesMap}</div>;
 };