+++ src/components/ElectoralCollege/ElectoralCollege.tsx
@@ -6,7 +6,7 @@ import States from '../States';
 import { STATES_DATA } from '../../constants';
 import { calculateEVPercentage } from '../../utilities';
 import '../../style.scss';
-import { IfcElectoralCollegeProps } from '../../types';
+import { IfcElectoralCollegeProps, StatesData, WinnerData } from '../../types';
 
 const checkForStickyEVCounter = () => {
   const candidatesEVCounter = document.querySelector('.candidatesWrapper');
@@ -26,12 +26,37 @@ const ElectoralCollege = (props: IfcElectoralCollegeProps) => {
   const {
     candidateImageSources,
     candidatesData,
+    dataMode = 'manual',
     enableStickyEVCounter = true,
+    handlePropError,
     mapSize = 'large',
+    proportionalReawardMode = 'stateWinner',
     stateControlSize = 'large',
+    voteTotals,
   } = props;
+  const renderErrorMessage = () => (
+    <>
+      This application is currently unable to display the Electoral College
+      widget.
+    </>
+  );
+  if (!candidatesData) {
+    handlePropError &&
+      handlePropError('candidatesData', 'candidatesData prop is required');
+    return renderErrorMessage();
+  }
+  if (dataMode === 'auto' && voteTotals?.length !== 51) {
+    handlePropError &&
+      handlePropError(
+        'voteTotals',
+        'voteTotals prop is required in auto mode and must contain data for all 50 states plus Washington, D.C. Do not include data for the Congressional Districts of Maine and Nebraska'
+      );
+    return renderErrorMessage();
+  }
+  const [hasPropError, setHasPropError] = useState<boolean>(false);
   const [isFromStorage, setIsFromStorage] = useState<boolean>(
-    Boolean(localStorage?.getItem('ElectoralCollegeStatus'))
+    dataMode === 'manual' &&
+      Boolean(localStorage?.getItem('ElectoralCollegeStatus'))
   );
   const startingEVTotals = isFromStorage
     ? JSON.parse(localStorage?.getItem('WinnerTakeAllTotals') as string)
@@ -72,12 +97,46 @@ const ElectoralCollege = (props: IfcElectoralCollegeProps) => {
   const [popVotesData, setPopVotesData] = useState(
     isFromStorage
       ? JSON.parse(localStorage.getItem('PopularVoteStatus') as string)
-      : []
+      : dataMode === 'auto'
+        ? voteTotals
+        : []
   );
+  const [hasCalculatedAllEVsFromPVs, setHasCalculatedAllEVsFromPVs] =
+    useState<boolean>(false);
+  const cumulativePVTotals = [0, 0, 0, 0, 0];
+  const cumulativeEVTotalsFromPVs = [0, 0, 0, 0, 0];
+  const statesDataCopy = [...statesData];
 
   useEffect(() => {
     if (enableStickyEVCounter) {
       window.onscroll = () => checkForStickyEVCounter();
+    }
+
+    if (dataMode === 'auto' && !hasCalculatedAllEVsFromPVs) {
+      setHasCalculatedAllEVsFromPVs(true);
+      const gopEvPct = calculateEVPercentage(cumulativeEVTotalsFromPVs[0]);
+      const demEvPct = calculateEVPercentage(cumulativeEVTotalsFromPVs[1]);
+      const libEvPct = calculateEVPercentage(cumulativeEVTotalsFromPVs[2]);
+      const grnEvPct = calculateEVPercentage(cumulativeEVTotalsFromPVs[3]);
+      const indEvPct = calculateEVPercentage(cumulativeEVTotalsFromPVs[4]);
+
+      const gopPvPct = calculateEVPercentage(cumulativePVTotals[0]);
+      const demPvPct = calculateEVPercentage(cumulativePVTotals[1]);
+      const libPvPct = calculateEVPercentage(cumulativePVTotals[2]);
+      const grnPvPct = calculateEVPercentage(cumulativePVTotals[3]);
+      const indPvPct = calculateEVPercentage(cumulativePVTotals[4]);
+
+      setPopularVoteTotals(cumulativePVTotals);
+      setPvPct([gopPvPct, demPvPct, libPvPct, grnPvPct, indPvPct]);
+
+      setGopTotal(cumulativeEVTotalsFromPVs[0]);
+      setDemTotal(cumulativeEVTotalsFromPVs[1]);
+      setLibTotal(cumulativeEVTotalsFromPVs[2]);
+      setGrnTotal(cumulativeEVTotalsFromPVs[3]);
+      setIndTotal(cumulativeEVTotalsFromPVs[4]);
+      setWinnerTakeAllTotals(cumulativeEVTotalsFromPVs);
+      setEvPct([gopEvPct, demEvPct, libEvPct, grnEvPct, indEvPct]);
+      setStatesData(statesDataCopy);
     }
   });
 
@@ -89,6 +148,12 @@ const ElectoralCollege = (props: IfcElectoralCollegeProps) => {
     setShowSavedMessage(showSavedMessage);
   }, [showSavedMessage]);
 
+  useEffect(() => {
+    if (voteTotals) {
+      setPopVotesData(voteTotals);
+    }
+  }, [setPopVotesData, voteTotals]);
+
   const handlePropVotes = (newPVTotals: Array<number>) => {
     const gopPvPct = calculateEVPercentage(newPVTotals[0]);
     const demPvPct = calculateEVPercentage(newPVTotals[1]);
@@ -98,6 +163,14 @@ const ElectoralCollege = (props: IfcElectoralCollegeProps) => {
 
     setPopularVoteTotals(newPVTotals);
     setPvPct([gopPvPct, demPvPct, libPvPct, grnPvPct, indPvPct]);
+  };
+
+  const handlePropVotesInAutoMode = (newPVTotals: Array<number>) => {
+    cumulativePVTotals[0] += newPVTotals[0];
+    cumulativePVTotals[1] += newPVTotals[1];
+    cumulativePVTotals[2] += newPVTotals[2];
+    cumulativePVTotals[3] += newPVTotals[3];
+    cumulativePVTotals[4] += newPVTotals[4];
   };
 
   const handleStateWinner = (newEVTotals: Array<number>) => {
@@ -114,6 +187,28 @@ const ElectoralCollege = (props: IfcElectoralCollegeProps) => {
     setIndTotal(newEVTotals[4]);
     setWinnerTakeAllTotals(newEVTotals);
     setEvPct([gopPct, demPct, libPct, grnPct, indPct]);
+  };
+
+  const handleStateWinnerInAutoMode = (
+    newEVTotals: Array<number>,
+    data?: WinnerData
+  ) => {
+    cumulativeEVTotalsFromPVs[0] += newEVTotals[0];
+    cumulativeEVTotalsFromPVs[1] += newEVTotals[1];
+    cumulativeEVTotalsFromPVs[2] += newEVTotals[2];
+    cumulativeEVTotalsFromPVs[3] += newEVTotals[3];
+    cumulativeEVTotalsFromPVs[4] += newEVTotals[4];
+
+    const stateId = data?.stateId;
+    const stateDataIndex = statesDataCopy.findIndex(
+      (state) => state.stateCode === stateId
+    );
+    // typecasting because we know it will be defined
+    const stateData = statesDataCopy.find(
+      (state) => state.stateCode === stateId
+    ) as StatesData;
+    stateData.winner = data?.newWinningParty;
+    statesDataCopy.splice(stateDataIndex, 1, stateData);
   };
 
   const resetWidget = () => {
@@ -145,34 +240,52 @@ const ElectoralCollege = (props: IfcElectoralCollegeProps) => {
       JSON.parse(localStorage.getItem('ElectoralCollegeStatus') as string)
     );
     setIsFromStorage(true);
-    setShowClearedMessage(false);
-    setShowSavedMessage(true);
-    setTimeout(() => setShowSavedMessage(false), 3000);
+    if (dataMode === 'manual') {
+      setShowClearedMessage(false);
+      setShowSavedMessage(true);
+      setTimeout(() => setShowSavedMessage(false), 3000);
+    }
   };
 
-  return (
-    <div className="electoral-college-root">
+  const handleSubComponentPropError = (prop: string, errorMessage: string) => {
+    handlePropError && handlePropError(prop, errorMessage);
+    setHasPropError(true);
+  };
+
+  return hasPropError ? (
+    renderErrorMessage()
+  ) : (
+    <div className={`electoral-college-root data-mode-${dataMode}`}>
       <CandidatesWrapper
         candidateImageSources={candidateImageSources}
         candidates={candidatesData}
         evPct={evPct}
         pvPct={pvPct}
         popVoteTotals={popularVoteTotals}
+        renderPropErrorMessage={handleSubComponentPropError}
         winnerTakeAllTotals={winnerTakeAllTotals}
       />
       <States
         currentEVTotals={winnerTakeAllTotals}
         currentPVTotals={popularVoteTotals}
-        handlePropVotes={handlePropVotes}
-        handleStateWinner={handleStateWinner}
+        dataMode={dataMode}
+        handlePropVotes={
+          dataMode === 'manual' ? handlePropVotes : handlePropVotesInAutoMode
+        }
+        handleStateWinner={
+          dataMode === 'manual'
+            ? handleStateWinner
+            : handleStateWinnerInAutoMode
+        }
         hasClearedSavedData={hasClearedSavedData}
         isFromStorage={isFromStorage}
         mapSize={mapSize}
         popVotesData={popVotesData}
+        proportionalReawardMode={proportionalReawardMode}
         stateControlSize={stateControlSize}
         statesData={statesData}
       />
-      {typeof Storage !== 'undefined' && (
+      {typeof Storage !== 'undefined' && dataMode === 'manual' && (
         <div id="buttons">
           <SaveButton
             currentEVTotals={winnerTakeAllTotals}