+++ src/utilities/__tests__/checkForWinner.test.ts
@@ -1,4 +1,4 @@
-import { checkForWinner } from '../';
+import { checkForWinner } from '..';
 
 const hasWinner = [55, 45, 0, 0, 0];
 const doesNotHaveWinner = [48, 49, 2, 0.5, 0.5];