+++ src/utilities/calculateStateEVs.ts
@@ -1,0 +1,4 @@
+const calculateStateEVs = (stateEVs: number, convertedVoteTotals: number[]) =>
+  convertedVoteTotals.map((votes) => Math.round((votes / 100) * stateEVs));
+
+export default calculateStateEVs;