+++ src/utilities/index.ts
@@ -1,13 +1,25 @@
-import calculateEVPercentage from './calculatePercentage';
+import calculateEVPercentage from './calculateEVPercentage';
+import calculateStateEVs from './calculateStateEVs';
+import calculateStateEVsFromPVs from './calculateStateEVsFromPVs';
 import checkForWinner from './checkForWinner';
 import getStateWinnerNames from './getStateWinnerNames';
+import getOrderedVoteTotals from './getOrderedVoteTotals';
 import getWinnerClassName from './getWinnerClassName';
+import getWinnerFromParty from './getWinnerFromParty';
 import getWinnerName from './getWinnerName';
+import getWinningPartyFullName from './getWinningParty';
+import reawardEVs from './reawardEVs';
 
 export {
   calculateEVPercentage,
+  calculateStateEVs,
+  calculateStateEVsFromPVs,
   checkForWinner,
+  getOrderedVoteTotals,
   getStateWinnerNames,
   getWinnerClassName,
+  getWinnerFromParty,
   getWinnerName,
+  getWinningPartyFullName,
+  reawardEVs,
 };